+++ _Typings/gameApi/types/player.d.ts
@@ -5,10 +5,6 @@ declare module "game:type" {
     import { AssertPath, AssertPaths } from "path"
     import { Float3, Quaternion } from "basicData"
 
-    /**
-     * Ball type can be any string if custom balls are considered.
-     * This type does not consider custom balls for easy typing.
-     */
     type BallType =
         | "WoodenBall"
         | "StoneBall"
@@ -19,6 +15,7 @@ declare module "game:type" {
         | "BalloonBall"
         | "StickyBall"
         | "SpongeBall"
+        | (string & {})
 
     /**
      * If the player ball is never created, errors will occur when trying to access it's properties other than `ballType` and `guid`.
@@ -26,10 +23,6 @@ declare module "game:type" {
     type Player = {
         /**
          * If the player ball is never created, the `ballType` will be `null`.
-         *
-         * @Note -
-         * - Ball type can be any string if custom balls are considered.
-         * - Custom balls are not considered here for easy typing.
          */
         get ballType(): BallType | null
         set ballType(ballType: BallType)